requirements.txt
matplotlib==3.7.2
numpy==1.23.5
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
